requirements.txt
Flask==2.3.3 
psycopg2-binary==2.9.9 
python-docx==1.1.0 
nltk==3.8.1 
tensorflow==2.15.0 
pandas==2.2.2 
numpy==1.26.4 
openai==1.30.0